tests/requirements-test.txt
# Core testing framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-mock>=3.12.0

# Docker integration
pytest-docker>=3.1.0
docker>=7.0.0

# Async support
asyncio>=3.4.3
aiofiles>=23.2.0

# Redis testing
redis>=5.0.0
fakeredis>=2.21.0  # For unit tests without real Redis

# Qdrant testing
qdrant-client>=1.7.0

# MinIO/S3 testing
minio>=7.2.0
boto3>=1.34.0

# MLflow testing
mlflow>=2.10.0

# HTTP/API testing
httpx>=0.26.0
requests>=2.31.0

# Data validation
pydantic>=2.5.0

# Test data generation
faker>=22.0.0
factory-boy>=3.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Code quality
black>=24.0.0
isort>=5.13.0
ruff>=0.1.0
mypy>=1.8.0

# Buildingblocks dependencies (from existing code)
celery>=5.3.0
kombu>=5.3.0
sqlalchemy>=2.0.0
fastapi>=0.109.0
pydantic-settings>=2.1.0
